+++ js/carousel.js
@@ -27,8 +27,6 @@
     constructor(el, options) {
       super(Carousel, el, options);
 
-      this.el = el;
-      this.$el = $(el);
       this.el.M_Carousel = this;
 
       /**
@@ -644,9 +642,9 @@
       if (index > this.count || index < 0) {
         if (this.noWrap) {
           return;
-        } else {
-          index = this._wrap(index);
         }
+
+        index = this._wrap(index);
       }
       this._cycleTo(index);
     }
@@ -664,9 +662,9 @@
       if (index > this.count || index < 0) {
         if (this.noWrap) {
           return;
-        } else {
-          index = this._wrap(index);
         }
+
+        index = this._wrap(index);
       }
 
       this._cycleTo(index);
@@ -685,9 +683,9 @@
       if (n > this.count || n < 0) {
         if (this.noWrap) {
           return;
-        } else {
-          n = this._wrap(n);
         }
+
+        n = this._wrap(n);
       }
 
       this._cycleTo(n, callback);