+++ js/datepicker.js
@@ -62,8 +62,8 @@
       nextMonth     : '›',
       months        : ['January','February','March','April','May','June','July','August','September','October','November','December'],
       monthsShort   : ['Jan','Feb','Mar','Apr','May','Jun','Jul','Aug','Sep','Oct','Nov','Dec'],
-      weekdaysShort : ['Sun','Mon','Tue','Wed','Thu','Fri','Sat'],
       weekdays      : ['Sunday','Monday','Tuesday','Wednesday','Thursday','Friday','Saturday'],
+      weekdaysShort : ['Sun','Mon','Tue','Wed','Thu','Fri','Sat'],
       weekdaysAbbrev : ['S','M','T','W','T','F','S']
     },
 
@@ -92,8 +92,6 @@
     constructor(el, options) {
       super(Datepicker, el, options);
 
-      this.el = el;
-      this.$el = $(el);
       this.el.M_Datepicker = this;
 
       this.options = $.extend({}, Datepicker.defaults, options);
@@ -188,7 +186,10 @@
      * Teardown component
      */
     destroy() {
-
+      this._removeEventHandlers();
+      this.modal.destroy();
+      $(this.modalEl).remove();
+      this.el.M_Datepicker = undefined;
     }
 
     _insertHTMLIntoDOM() {
@@ -236,9 +237,9 @@
       let formattedDate = formatArray.map((label) => {
         if (this.formats[label]) {
           return this.formats[label]();
-        } else {
-          return label;
         }
+
+        return label;
       }).join( '' );
       return formattedDate;
     }
@@ -620,8 +621,8 @@
       // Init Materialize Select
       let yearSelect = this.calendarEl.querySelector('.pika-select-year');
       let monthSelect = this.calendarEl.querySelector('.pika-select-month');
-      M.Select.init(yearSelect, {classes: 'select-year'});
-      M.Select.init(monthSelect, {classes: 'select-month'});
+      M.Select.init(yearSelect, {classes: 'select-year', dropdownOptions: {container: document.body, constrainWidth: false}});
+      M.Select.init(monthSelect, {classes: 'select-month', dropdownOptions: {container: document.body, constrainWidth: false}});
 
       // Add change handlers for select
       yearSelect.addEventListener('change', this._handleYearChange.bind(this));
@@ -673,8 +674,12 @@
 
       this.formats = {
 
-        dd: () => {
+        d: () => {
           return this.date.getDate();
+        },
+        dd: () => {
+          let d = this.date.getDate();
+          return (d < 10 ? '0' : '') + d;
         },
         ddd: () => {
           return this.options.i18n.weekdaysShort[this.date.getDay()];
@@ -682,17 +687,21 @@
         dddd: () => {
           return this.options.i18n.weekdays[this.date.getDay()];
         },
-        mm: () => {
+        m: () => {
           return this.date.getMonth() + 1;
+        },
+        mm: () => {
+          let m = this.date.getMonth() + 1;
+          return (m < 10 ? '0' : '') + m;
         },
         mmm: () => {
           return this.options.i18n.monthsShort[this.date.getMonth()];
         },
         mmmm: () => {
-          return this.options.i18n.monthsShort[this.date.getMonth()];
+          return this.options.i18n.months[this.date.getMonth()];
         },
         yy: () => {
-          return this.date.getFullYear().slice(2);
+          return ('' + this.date.getFullYear()).slice(2);
         },
         yyyy: () => {
           return this.date.getFullYear();