+++ js/global.js
@@ -68,24 +68,23 @@ M.initializeJqueryWrapper = function(plugin, pluginName, classRef) {
       if (methodOrOptions.slice(0,3) === 'get') {
         let instance = this.first()[0][classRef];
         return instance[methodOrOptions].apply(instance, params);
+      }
 
       // Void methods
-      } else {
-        return this.each(function() {
-          let instance = this[classRef];
-          instance[methodOrOptions].apply(instance, params);
-        });
-      }
+      return this.each(function() {
+        let instance = this[classRef];
+        instance[methodOrOptions].apply(instance, params);
+      });
 
     // Initialize plugin if options or no argument is passed in
     } else if ( typeof methodOrOptions === 'object' || ! methodOrOptions ) {
       plugin.init(this, arguments[0]);
       return this;
 
-    // Return error if an unrecognized  method name is passed in
-    } else {
-      jQuery.error(`Method ${methodOrOptions} does not exist on jQuery.${pluginName}`);
     }
+
+    // Return error if an unrecognized  method name is passed in
+    jQuery.error(`Method ${methodOrOptions} does not exist on jQuery.${pluginName}`);
   };
 };
 
@@ -121,7 +120,7 @@ M.guid = (function() {
  * @returns {string}
  */
 M.escapeHash = function(hash) {
-  return hash.replace( /(:|\.|\[|\]|,|=)/g, "\\$1" );
+  return hash.replace( /(:|\.|\[|\]|,|=|\/)/g, "\\$1" );
 };
 
 M.elementOrParentIsFixed = function(element) {
@@ -215,43 +214,42 @@ M.checkPossibleAlignments = function(el, container, bounding, offset) {
 
   let containerAllowsOverflow = getComputedStyle(container).overflow === 'visible';
   let containerRect = container.getBoundingClientRect();
+  let containerHeight = Math.min(containerRect.height, window.innerHeight);
+  let containerWidth = Math.min(containerRect.width, window.innerWidth);
   let elOffsetRect = el.getBoundingClientRect();
 
   let scrollLeft = container.scrollLeft;
   let scrollTop = container.scrollTop;
 
   let scrolledX = bounding.left - scrollLeft;
-  let scrolledY = bounding.top - scrollTop;
+  let scrolledYTopEdge = bounding.top - scrollTop;
+  let scrolledYBottomEdge = bounding.top + elOffsetRect.height - scrollTop;
 
   // Check for container and viewport for left
-  canAlign.spaceOnRight = !containerAllowsOverflow ? container.offsetWidth - (scrolledX + bounding.width) :
+  canAlign.spaceOnRight = !containerAllowsOverflow ? containerWidth - (scrolledX + bounding.width) :
     window.innerWidth - (elOffsetRect.left + bounding.width);
-  if ((!containerAllowsOverflow && scrolledX + bounding.width > container.offsetWidth) ||
-      containerAllowsOverflow && (elOffsetRect.left + bounding.width > window.innerWidth)) {
+  if (canAlign.spaceOnRight < 0) {
     canAlign.left = false;
   }
 
   // Check for container and viewport for Right
   canAlign.spaceOnLeft = !containerAllowsOverflow ? scrolledX - bounding.width + elOffsetRect.width :
     elOffsetRect.right - bounding.width;
-  if ((!containerAllowsOverflow && scrolledX - bounding.width + elOffsetRect.width < 0) ||
-      containerAllowsOverflow && (elOffsetRect.right - bounding.width < 0)) {
+  if (canAlign.spaceOnLeft < 0) {
     canAlign.right = false;
   }
 
   // Check for container and viewport for Top
-  canAlign.spaceOnBottom = !containerAllowsOverflow ? containerRect.height - (scrolledY + bounding.height + offset) :
+  canAlign.spaceOnBottom = !containerAllowsOverflow ? containerHeight - (scrolledYTopEdge + bounding.height + offset) :
     window.innerHeight - (elOffsetRect.top + bounding.height + offset);
-  if ((!containerAllowsOverflow && scrolledY + bounding.height + offset > containerRect.height) ||
-      containerAllowsOverflow && (elOffsetRect.top + bounding.height + offset > window.innerHeight)) {
+  if (canAlign.spaceOnBottom < 0) {
     canAlign.top = false;
   }
 
   // Check for container and viewport for Bottom
-  canAlign.spaceOnTop = !containerAllowsOverflow ? scrolledY - (bounding.height + offset) :
+  canAlign.spaceOnTop = !containerAllowsOverflow ? scrolledYBottomEdge - (bounding.height - offset) :
     elOffsetRect.bottom - (bounding.height + offset);
-  if ((!containerAllowsOverflow && scrolledY - bounding.height - offset < 0) ||
-      containerAllowsOverflow && (elOffsetRect.bottom - bounding.height - offset < 0)) {
+  if (canAlign.spaceOnTop < 0) {
     canAlign.bottom = false;
   }
 
@@ -266,9 +264,9 @@ M.getOverflowParent = function(element) {
 
   if (element === document.body || getComputedStyle(element).overflow !== 'visible') {
     return element;
-  } else {
-    return M.getOverflowParent(element.parentElement);
   }
+
+  return M.getOverflowParent(element.parentElement);
 };