+++ js/parallax.js
@@ -2,20 +2,30 @@
   'use strict';
 
   let _defaults = {
+    responsiveThreshold: 0, // breakpoint for swipeable
   };
 
-  class Parallax extends Component{
+  class Parallax extends Component {
 
     constructor(el, options) {
       super(Parallax, el, options);
 
-      this.el = el;
-      this.$el = $(el);
       this.el.M_Parallax = this;
 
+      /**
+       * Options for the Parallax
+       * @member Parallax#options
+       * @prop {Number} responsiveThreshold
+       */
       this.options = $.extend({}, Parallax.defaults, options);
+      this._enabled = window.innerWidth > this.options.responsiveThreshold;
 
       this.$img = this.$el.find('img').first();
+      this.$img.each(function() {
+        let el = this;
+        if (el.complete) $(el).trigger("load");
+      });
+
       this._updateParallax();
       this._setupEventHandlers();
       this._setupStyles();
@@ -43,13 +53,24 @@
      * Teardown component
      */
     destroy() {
+      Parallax._parallaxes.splice(Parallax._parallaxes.indexOf(this), 1);
+      this.$img[0].style.transform = '';
+      this._removeEventHandlers();
 
+      this.$el[0].M_Parallax = undefined;
     }
 
     static _handleScroll() {
-      for(let i = 0; i < Parallax._parallaxes.length; i++) {
+      for (let i = 0; i < Parallax._parallaxes.length; i++) {
         let parallaxInstance = Parallax._parallaxes[i];
         parallaxInstance._updateParallax.call(parallaxInstance);
+      }
+    }
+
+    static _handleWindowResize() {
+      for (let i = 0; i < Parallax._parallaxes.length; i++) {
+        let parallaxInstance = Parallax._parallaxes[i];
+        parallaxInstance._enabled = window.innerWidth > parallaxInstance.options.responsiveThreshold;
       }
     }
 
@@ -60,6 +81,18 @@
       if (Parallax._parallaxes.length === 0) {
         Parallax._handleScrollThrottled = M.throttle(Parallax._handleScroll, 5);
         window.addEventListener('scroll', Parallax._handleScrollThrottled);
+
+        Parallax._handleWindowResizeThrottled = M.throttle(Parallax._handleWindowResize, 5);
+        window.addEventListener('resize', Parallax._handleWindowResizeThrottled);
+      }
+    }
+
+    _removeEventHandlers() {
+      this.$img[0].removeEventListener('load', this._handleImageLoadBound);
+
+      if (Parallax._parallaxes.length === 0) {
+        window.removeEventListener('scroll', Parallax._handleScrollThrottled);
+        window.removeEventListener('resize', Parallax._handleWindowResizeThrottled);
       }
     }
 
@@ -69,10 +102,6 @@
 
     _handleImageLoad() {
       this._updateParallax();
-      this.$img.each(function() {
-        let el = this;
-        if (el.complete) $(el).trigger("load");
-      });
     }
 
     _updateParallax() {
@@ -87,7 +116,10 @@
       let percentScrolled = (windowBottom - top) / (containerHeight + windowHeight);
       let parallax = parallaxDist * percentScrolled;
 
-      if (bottom > scrollTop && top < scrollTop + windowHeight) {
+      if (!this._enabled) {
+        this.$img[0].style.transform = '';
+
+      } else if (bottom > scrollTop && top < scrollTop + windowHeight) {
         this.$img[0].style.transform = `translate3D(-50%, ${parallax}px, 0)`;
       }
     }