+++ js/select.js
@@ -2,7 +2,8 @@
   'use strict';
 
   let _defaults = {
-    classes: ''
+    classes: '',
+    dropdownOptions: {}
   };
 
 
@@ -20,8 +21,11 @@
     constructor(el, options) {
       super(Select, el, options);
 
-      this.el = el;
-      this.$el = $(el);
+      // Don't init if browser default version
+      if (this.$el.hasClass('browser-default')) {
+        return;
+      }
+
       this.el.M_Select = this;
 
       /**
@@ -33,6 +37,7 @@
       this.isMultiple = this.$el.prop('multiple');
 
       // Setup
+      this.el.tabIndex = -1;
       this._keysSelected = {};
       this._valueDict = {}; // Maps key to original and generated option element.
       this._setupDropdown();
@@ -89,7 +94,6 @@
       });
       this.el.removeEventListener('change', this._handleSelectChangeBound);
       this.input.removeEventListener('click', this._handleInputClickBound);
-      this.input.removeEventListener('focus', this._handleInputFocusBound);
     }
 
     /**
@@ -216,11 +220,10 @@
 
       // Initialize dropdown
       if (!this.el.disabled) {
-        let dropdownOptions = {};
         if (this.isMultiple) {
-          dropdownOptions.closeOnClick = false;
+          this.options.dropdownOptions.closeOnClick = false;
         }
-        this.dropdown = M.Dropdown.init(this.input, dropdownOptions);
+        this.dropdown = M.Dropdown.init(this.input, this.options.dropdownOptions);
       }
 
       // Add initial selections