+++ js/timepicker.js
@@ -10,9 +10,15 @@
     container: null,
     defaultTime: 'now',         // default time, 'now' or '13:14' e.g.
 		fromnow: 0,            // Millisecond offset from the defaultTime
-		doneText: 'Ok',      // done button text
-		clearText: 'Clear',
-		cancelText: 'Cancel',
+
+    clearBtn: true,
+    // internationalization
+    i18n: {
+      done: 'Ok',
+		  clear: 'Clear',
+		  cancel: 'Cancel',
+    },
+
 		autoClose: false,      // auto close when minute is selected
 		twelveHour: true,      // change to 12 hour AM/PM clock from 24 hour
 		vibrate: true          // vibrate the device when dragging clock hand
@@ -27,8 +33,6 @@
     constructor(el, options) {
       super(Timepicker, el, options);
 
-      this.el = el;
-      this.$el = $(el);
       this.el.M_Timepicker = this;
 
       this.options = $.extend({}, Timepicker.defaults, options);
@@ -232,17 +236,17 @@
     }
 
     _pickerSetup() {
-        if (this.options.clearText) {
-            $('<button class="btn-flat timepicker-clear waves-effect" type="button" tabindex="' + (this.options.twelveHour? '3' : '1') + '">' + this.options.clearText + '</button>')
-            .appendTo(this.footer).on('click', this.clear.bind(this));
+        if (this.options.clearBtn) {
+            $('<button class="btn-flat timepicker-clear waves-effect" type="button" tabindex="' + (this.options.twelveHour? '3' : '1') + '">' + this.options.i18n.clear + '</button>')
+                .appendTo(this.footer).on('click', this.clear.bind(this));
         } else {
             $(this.footer).append('<span></span>'); // push  buttons to the right
         }
 
       let confirmationBtnsContainer = $('<div class="confirmation-btns"></div>');
-		  $('<button class="btn-flat timepicker-close waves-effect" type="button" tabindex="' + (this.options.twelveHour? '3' : '1') + '">' + this.options.cancelText + '</button>')
+		  $('<button class="btn-flat timepicker-close waves-effect" type="button" tabindex="' + (this.options.twelveHour? '3' : '1') + '">' + this.options.i18n.cancel + '</button>')
         .appendTo(confirmationBtnsContainer).on('click', this.close.bind(this));
-		  $('<button class="btn-flat timepicker-close waves-effect" type="button" tabindex="' + (this.options.twelveHour? '3' : '1') + '">' + this.options.doneText + '</button>')
+		  $('<button class="btn-flat timepicker-close waves-effect" type="button" tabindex="' + (this.options.twelveHour? '3' : '1') + '">' + this.options.i18n.done + '</button>')
         .appendTo(confirmationBtnsContainer).on('click', this.done.bind(this));
       confirmationBtnsContainer.appendTo(this.footer);
     }
@@ -361,7 +365,7 @@
       // Get the time
 		  let value = ((this.el.value || this.options.defaultTime || '') + '').split(':');
 		  if (this.options.twelveHour && !(typeof value[1] === 'undefined')) {
-			  if (value[1].indexOf("AM") > 0){
+			  if (value[1].toUpperCase().indexOf("AM") > 0){
 				  this.amOrPm = 'AM';
 			  } else {
 				  this.amOrPm = 'PM';