+++ js/tooltip.js
@@ -27,8 +27,6 @@
     constructor(el, options) {
       super(Tooltip, el, options);
 
-      this.el = el;
-      this.$el = $(el);
       this.el.M_Tooltip = this;
       this.options = $.extend({}, Tooltip.defaults, options);
 
@@ -121,9 +119,9 @@
       this._exitDelayTimeout = setTimeout(() => {
         if (this.isHovered) {
           return;
-        } else {
-          this._animateOut();
         }
+
+        this._animateOut();
       }, this.options.exitDelay);
     }
 
@@ -136,9 +134,9 @@
       this._enterDelayTimeout = setTimeout(() => {
         if (!this.isHovered) {
           return;
-        } else {
-          this._animateIn();
         }
+
+        this._animateIn();
       }, this.options.enterDelay);
     }