+++ CHANGELOG.md
@@ -2,9 +2,14 @@ Changelog
 =======
 Bolded styling surrounded by emojis indicates a breaking change.
 
-## 1.0.0-alpha.3 (TBD)
-- *Initialization code for components changed. E.g. Change`new M.Tooltip(el, options)` to `M.Tooltip.init(el, options)`*
+## 1.0.0-alpha.3 (December 29th, 2017)
+- *Initialization code for all components changed. E.g. Change`new M.Tooltip(el, options)` to `M.Tooltip.init(el, options)`*
   - This was done so that the same initialization code can be used to initialize single Elements as well as NodeLists and jQuery element objects
+
+- Added in onOpen and OnClose callbacks for appropriate plugins
+
+- Datepicker
+  - Destroy function added
 
 - Feature Discovery
   - Added open and close callbacks
@@ -15,6 +20,10 @@ Bolded styling surrounded by emojis indicates a breaking change.
 - Modal
   - Added open and close callbacks for consistency
   - Removed ready and complete callbacks
+
+- Parallax
+  - Has responsiveThreshold option
+  - Destroy function added
 
 - Pushpin
   - added onPositionChange callback